fix(auctions): address CodeRabbit review findings for type safety and error handling
