+++ conductor/tracks/refactor_auction_mutations_20260321/index.md
@@ -1,0 +1,5 @@
+# Track refactor_auction_mutations_20260321 Context
+
+- [Specification](./spec.md)
+- [Implementation Plan](./plan.md)
+- [Metadata](./metadata.json)
