+++ convex/_generated/api.d.ts
@@ -18,11 +18,16 @@ import type * as admin_statistics from "../admin/statistics.js";
 import type * as admin_debug from "../admin_debug.js";
 import type * as admin_utils from "../admin_utils.js";
 import type * as auctions from "../auctions.js";
-import type * as auctions_bidding from "../auctions/bidding.js";
 import type * as auctions_helpers from "../auctions/helpers.js";
 import type * as auctions_index from "../auctions/index.js";
 import type * as auctions_internal from "../auctions/internal.js";
-import type * as auctions_mutations from "../auctions/mutations.js";
+import type * as auctions_mutations_bidding from "../auctions/mutations/bidding.js";
+import type * as auctions_mutations_create from "../auctions/mutations/create.js";
+import type * as auctions_mutations_delete from "../auctions/mutations/delete.js";
+import type * as auctions_mutations_helpers from "../auctions/mutations/helpers.js";
+import type * as auctions_mutations_index from "../auctions/mutations/index.js";
+import type * as auctions_mutations_publish from "../auctions/mutations/publish.js";
+import type * as auctions_mutations_update from "../auctions/mutations/update.js";
 import type * as auctions_proxy_bidding from "../auctions/proxy_bidding.js";
 import type * as auctions_queries from "../auctions/queries.js";
 import type * as auctions_queries_admin from "../auctions/queries/admin.js";
@@ -64,11 +69,16 @@ declare const fullApi: ApiFromModules<{
   admin_debug: typeof admin_debug;
   admin_utils: typeof admin_utils;
   auctions: typeof auctions;
-  "auctions/bidding": typeof auctions_bidding;
   "auctions/helpers": typeof auctions_helpers;
   "auctions/index": typeof auctions_index;
   "auctions/internal": typeof auctions_internal;
-  "auctions/mutations": typeof auctions_mutations;
+  "auctions/mutations/bidding": typeof auctions_mutations_bidding;
+  "auctions/mutations/create": typeof auctions_mutations_create;
+  "auctions/mutations/delete": typeof auctions_mutations_delete;
+  "auctions/mutations/helpers": typeof auctions_mutations_helpers;
+  "auctions/mutations/index": typeof auctions_mutations_index;
+  "auctions/mutations/publish": typeof auctions_mutations_publish;
+  "auctions/mutations/update": typeof auctions_mutations_update;
   "auctions/proxy_bidding": typeof auctions_proxy_bidding;
   "auctions/queries": typeof auctions_queries;
   "auctions/queries/admin": typeof auctions_queries_admin;
