+++ convex/auctions/index.ts
@@ -1,6 +1,6 @@
 export * from "./queries";
 export * from "./mutations";
-export * from "./bidding";
+export * from "./mutations/bidding";
 export * from "./internal";
 export * from "./helpers";
 export {
