+++ src/components/auction/AuctionCard.tsx
@@ -39,7 +39,7 @@ export const AuctionCard = ({
 }: AuctionCardProps) => {
   const { data: session } = useSession();
   const navigate = useNavigate();
-  const placeBid = useMutation(api.auctions.placeBid);
+  const placeBid = useMutation(api.auctions.mutations.bidding.placeBid);
   const toggleWatchlist = useMutation(api.watchlist.toggleWatchlist);
   const [isBidding, setIsBidding] = useState(false);
   const isBiddingRef = useRef(false);
